+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 
 
+## 2026-07-10
+
+### Features
+
+* send push notifications for likes, comments, follows, and recommendations ([f92a75f](https://adl.github.com/adl-developer/kinkane-backend/commit/f92a75f30ac8c5a7088e67729ca8ca89ae6904ab))
+
+
 ## 2026-07-09
 
 ### Features